Refactor: libcrmcommon: pcmk__compress now takes size_t arguments.
The idea here is that we can consolidate some of the integer overflow
checking into this function, making its callers less complicated.

Note that send_cpg_text needs a lot of the same checking that's been
added to pcmk__remote_send_xml recently, though we don't need to be
quite as defensive in that function.  We don't allow just any client to
connect to the cluster layer, so the potential for bad actors should be
considerably lower.

For the moment I'm just making sure the function continues to compile.
Also note that send_cpg_text never returns false indicating that sending
failed.  There's probably a lot of error handling that could be done.

+++ lib/common/tests/strings/pcmk__compress_test.c
@@ -23,7 +23,7 @@ static void
 simple_compress(void **state)
 {
     char *result = pcmk__assert_alloc(1024, sizeof(char));
-    unsigned int len;
+    size_t len;
 
     assert_int_equal(pcmk__compress(SIMPLE_DATA, 40, &result, &len), pcmk_rc_ok);
     assert_memory_equal(result, SIMPLE_COMPRESSED, 13);
@@ -32,7 +32,7 @@ simple_compress(void **state)
 static void
 calloc_fails(void **state) {
     char *result = pcmk__assert_alloc(1024, sizeof(char));
-    unsigned int len;
+    size_t len;
 
     pcmk__assert_exits(
         CRM_EX_OSERR,
